Fixed flags.

+++ philo/src/mutex_init.c
@@ -20,7 +20,7 @@ int	mutex_init(t_philo *philo)
 	philo->mutex_array = (pthread_mutex_t *) malloc (sizeof(pthread_mutex_t) * (philo->philo_count + 1));
 	if (!philo->mutex_array)
 		return (ERR_MEM_ALLOC_FAILED);
-	while (i < philo->philo_count)
+	while (i < (size_t)philo->philo_count)
 	{
 		if (pthread_mutex_init(&philo->mutex_array[i], NULL) != 0)
 			return (free(philo->mutex_array), ERR_MUTEX_FAILED);
